% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_fxns.R
\name{collect_fxns}
\alias{collect_fxns}
\title{Collect functions into an environment}
\usage{
collect_fxns(path)
}
\arguments{
\item{path}{(character) path to a package, default: current directory}
}
\value{
an enironment with functions in the environment
}
\description{
Collect functions into an environment
}
