% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fxns.R
\name{make_fxn}
\alias{make_fxn}
\alias{make_fxns}
\title{Make functions from \code{\link[utils:getParseData]{utils::getParseData()}} output}
\usage{
make_fxn(x)

make_fxns(x)
}
\arguments{
\item{x}{for \code{make_fxn}, a data.frame, as output from
\code{\link[utils:getParseData]{utils::getParseData()}}; for \code{make_fxns}, a list of those}
}
\value{
for \code{make_fxn} a function as a character string;
for \code{make_fxns}, a list of those
}
\description{
Make functions from \code{\link[utils:getParseData]{utils::getParseData()}} output
}
\examples{
\dontrun{
foo <- function(x) {
  if (x == 1) x else 5
}
foo
f <- mutate_one(foo)
f
x = f
make_fxn(f)

bar <- function(w) {
  if (w == 10) w else 5
}
g <- mutate_one(bar)
make_fxns(list(f, g))
}
}
\keyword{internal}
