% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pkg_map.R
\name{make_pkg_map}
\alias{make_pkg_map}
\title{Make a package map}
\usage{
make_pkg_map(path = ".")
}
\arguments{
\item{path}{(character) path to a package, default: current directory}
}
\value{
the output from \code{\link[pkgapi:map_package]{pkgapi::map_package()}}; a thin wrapper around
that fxn
}
\description{
Using \code{\link[pkgapi:map_package]{pkgapi::map_package()}}
}
