% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutant-package.R
\docType{package}
\name{mutant-package}
\alias{mutant}
\alias{mutant-package}
\title{mutant: Mutation Testing Framework}
\description{
Mutation testing framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sckott.github.io/mutant/}
  \item \url{https://github.com/sckott/mutant}
  \item Report bugs at \url{https://github.com/sckott/mutant/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{myrmecocystus@gmail.com} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})

}
\keyword{internal}
