% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_one}
\alias{mutate_one}
\alias{mutate_one.default}
\alias{mutate_one.function}
\alias{mutate_one.ast}
\title{Make a mutation in one function}
\usage{
mutate_one(x)

\method{mutate_one}{default}(x)

\method{mutate_one}{`function`}(x)

\method{mutate_one}{ast}(x)
}
\arguments{
\item{x}{a data.frame, the output of \code{\link[utils:getParseData]{utils::getParseData()}}, called
from \code{\link[=parse_fxns]{parse_fxns()}}}
}
\value{
the same data.frame as in \code{x}, but with a single mutation
}
\description{
Make a mutation in one function
}
\note{
uses \code{\link[astr:ast_modify]{astr::ast_modify()}} internally
}
\examples{
\dontrun{
foo <- function(x) {
  if (x == 1) x else 5
}
mutate_one(foo)
mutate_one(astr::ast_decompose(foo))
}
}
\keyword{internal}
