% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_test.R
\name{mutation_test}
\alias{mutation_test}
\title{Mutation test}
\usage{
mutation_test(x)
}
\arguments{
\item{x}{(character) complete path to the package}
}
\value{
the output of \link[testthat:SilentReporter]{testthat::SilentReporter}
}
\description{
Run \code{\link[testthat:test_dir]{testthat::test_dir()}} on a mutated package
}
