% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fxns.R
\name{parse_fxns}
\alias{parse_fxns}
\title{Parse functions}
\usage{
parse_fxns(env)
}
\arguments{
\item{env}{an environment with functions}
}
\value{
a list of data.frame's, each of which is the output from
\code{\link[utils:getParseData]{utils::getParseData()}}; names of the list are the function names
}
\description{
Parse functions
}
\examples{
\dontrun{
path <- "../randgeo/"
env <- collect_fxns(path)
parse_fxns(env)
}
}
\keyword{internal}
