% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_test}
\alias{run_test}
\alias{run_tests}
\title{Run tests a single time}
\usage{
run_test(map)

run_tests(map, n = 1)
}
\arguments{
\item{map}{a function map as returned from \code{\link[=make_pkg_map]{make_pkg_map()}}}

\item{n}{(integer) number of times to run tests}
}
\value{
an object of class \code{testthat_results}
}
\description{
Run tests a single time
}
\details{
uses \link[testthat:SilentReporter]{testthat::SilentReporter} as the reporter
in \code{\link[testthat:test_package]{testthat::test_package()}}
}
\examples{
\dontrun{
path <- '/Users/sckott/github/ropensci/rredlist'
pkgmap <- make_pkg_map(path)

# single test
res <- run_test(pkgmap)
class(res)
res[[1]]

# many tests
out <- run_tests(pkgmap, times = 2)
length(out)
out[[1]]
out[[2]]
}
}
\keyword{internal}
