% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mutated_pkg.R
\name{write_mutated_pkg}
\alias{write_mutated_pkg}
\title{Write mutated package}
\usage{
write_mutated_pkg(pkg_path, fxns, map)
}
\arguments{
\item{pkg_path}{(character) path to the original package. we don't overwrite
any files in the original path/location}

\item{fxns}{(list) list of functions, output of \code{\link[=make_fxns]{make_fxns()}}}

\item{map}{(list) package map, output of \code{\link[=make_pkg_map]{make_pkg_map()}}}
}
\description{
Write mutated package
}
\keyword{internal}
