% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutant-package.R
\docType{package}
\name{mutant-package}
\alias{mutant-package}
\alias{mutant}
\title{mutant}
\description{
Mutation Testing Framework
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
