% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutaters.R
\name{mutaters}
\alias{mutaters}
\title{mutaters}
\description{
mutaters

mutaters
}
\examples{
\dontrun{
x <- mutaters$new()
x
# select mutater by name
x$muts$bool_false2true
# fetch a random mutater
z <- x$random()
z
z$name
z$description
z$from
z$to
}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mutater}}{(character) a mutater name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{mutaters$print()}}
\item \href{#method-new}{\code{mutaters$new()}}
\item \href{#method-random}{\code{mutaters$random()}}
\item \href{#method-clone}{\code{mutaters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method for \code{mutaters} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mutaters$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new mutaters object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mutaters$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{mutaters} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random"></a>}}
\if{latex}{\out{\hypertarget{method-random}{}}}
\subsection{Method \code{random()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mutaters$random()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mutaters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
